insert file log & load to staging

+++ src/main/java/com/dao/IFileLogDao.java
@@ -1,7 +1,7 @@
 package com.dao;
 
 public interface IFileLogDao {
-    int insert(int configId, String path, String status, int authorId);
+    long insert(int configId, String path, String status, int authorId);
 
     boolean updateFileLog(int id,String name);
 }
