+++ src/main/java/com/service/IFileLogService.java
@@ -1,7 +1,7 @@
 package com.service;
 
 public interface IFileLogService {
-    int insert( int configId,String path, String status, int authorId);
+    long insert(int configId, String path, String status, int authorId);
 
     boolean updateFileLog(int id, String status);
